+++ gems/aws-partitions/CHANGELOG.md
@@ -3,6 +3,11 @@ Unreleased Changes
 
 * Feature - Add partition metadata module, allowing access without loading entire partitions.json.
 
+1.986.0 (2024-10-07)
+------------------
+
+* Feature - Updated the partitions source data the determines the AWS service regions and endpoints.
+
 1.985.0 (2024-10-04)
 ------------------