+++ gems/aws-sdk-marketplacereporting/CHANGELOG.md
@@ -1,6 +1,11 @@
 Unreleased Changes
 ------------------
 
+1.1.0 (2024-10-07)
+------------------
+
+* Feature - Documentation-only update for AWS Marketplace Reporting API.
+
 1.0.0 (2024-10-03)
 ------------------